Extract EtagPolling from EventsSource (#47)
